Compositions

JuLie.Composition — Type

Composition{T} <: AbstractArray{T,1}

A composition of an integer n ≥ 0 is a sequence (λ₁,…,λₖ) of positive integers λᵢ whose sum is equal to n. Compositions can also be thought of as ordered partitions of n.

Compositions are implemented as a subtype of 1-dimensional integer arrays and you can thus work with compositions like with arrays. You can use smaller integer types to increase performance.

Examples

julia> P=Composition([2,1])
[2, 1]
julia> sum(P)
3
julia> P[1]
2
julia> P=Composition(Int8[2,1])
Int8[2, 1]

References

Wikipedia, Composition (combinatorics)

Unrestricted compositions

JuLie.compositions — Method

compositions(n::Integer)

Returns an array of all compositions of an integer n. This iterates over compositions(n::Integer, k::Integer) of n into k parts for 1 ≤ k ≤ n.

Examples

julia> compositions(3)
4-element Vector{Composition{Int64}}:
 [3]
 [1, 2]
 [2, 1]
 [1, 1, 1]

JuLie.num_compositions — Method

num_compositions(n::Integer)

The number of compositions of an integer $n ≥ 0$ is equal to

\[\left\lbrace \begin{array}{ll} 1 & \text{if } n = 0 \\ 2^{n-1} & \text{if } n ≥ 1 \;. \end{array} \right.\]

See Corollary 5.4 of Bóna (2017).

References

The On-Line Encyclopedia of Integer Sequences, A011782

Bóna, M. (2017). A Walk Through Combinatorics (Fourth Edition). World Scientific.

Restricted compositions

JuLie.compositions — Method

compositions(n::Integer, k::Integer)

Returns an array of all compositions of n into k parts. The algorithm used is Algorithm 72 "Composition Generator" by Hellerman & Ogden (1961), which refers to Chapter 6 of Riordan (1958). De-gotoed by E. Thiel. I don't know if there are faster algorithms but this one is already very fast.

Examples

julia> compositions(4,2)
3-element Vector{Composition{Int64}}:
 [1, 3]
 [2, 2]
 [3, 1]

References

Hellerman, L. & Ogden, S. (1961). Algorithm 72: composition generator. Communications of the ACM, 4(11), 498. https://doi.org/10.1145/366813.366837

Riordan, J. (1958). An introduction to Combinatorial Analysis. John Wiley & Sons, Inc., New York.

JuLie.num_compositions — Method

num_compositions(n::Integer, k::Integer)

The number of compositions of an integer $n ≥ 0$ into $k ≥ 0$ parts is equal

\[\left\lbrace \begin{array}{ll} 1 & \text{if } n=k=0 \\ {n-1 \choose k-1} & \text{otherwise} \;. \end{array} \right.\]

See Corollary 5.3 of Bóna (2017).

References

Bóna, M. (2017). A Walk Through Combinatorics (Fourth Edition). World Scientific.