JuLie.jl

An early-stage Julia/OSCAR package for things around algebraic Lie theory, representation theory, and relevant combinatorics. Contributions are very much welcome!

By Ulrich Thiel (ulthiel.com/math) and contributors.

Using

To install the package, do the following In Julia:

using Pkg
Pkg.add(url="https://github.com/ulthiel/JuLie.jl")

Then you can start using the package as follows:

using JuLie
partitions(10)

You can get a list of exported functions using

names(JuLie)

You can get help for a function by putting a question mark in front, e.g.

?partitions

Partitions

JuLie.Partition — Type

struct Partition{T} <: AbstractArray{T,1}

A partition of an integer $n \geq 0$ is a decreasing (our convention) sequence $n_1, n_2, \ldots$ of positive integers whose sum is equal to $n$. You can create a partition with

P=Partition([3,2,1])

and then work with it like with an array. In fact, Partition is a subtype of AbstractArray{T,1}. You can increase performance by using smaller integer types, e.g.

P=Partition(Int8[3,2,1])

Note that for efficiency the Partition constructor does not check whether the given array is in fact a partition, i.e. a decreasing sequence. That's your job.

Remark. I was thinking back and forth whether to implement an own structure for this because it's actually just an array of integers. But it makes sense since we have several functions just acting on partitons and it would be strange implementing them for arrays in general (where mostly they don't make sense). I was hesitating because I feared that an own structure for partitions will have a performance impact. But it does not! In my standard example creating the partitions of 90 there is really NO difference in runtime and memory consumption between using arrays and using an own structure.

The implementation of a subtype of AbstractArray is explained in the Julia documentation.

JuLie.getelement — Function

getelement(P::Partition, i::Int)

returns the $i$-th Element of $P$, if $i > length(P)$ this returns 0 instead of throwing an Exception.

If you are sure that $P[i]$ exists, use getindex which is faster.

JuLie.dominates — Function

dominates(lambda::Partition, mu::Partition)

returns true if $lambda ≥ mu$ according to the dominance order on partitions:

$λ ≥ μ :⟺ λ[1] + ... + λ[i] ≥ μ[1] + ... + μ[i]$ for all $i$

JuLie.conjugate — Function

conjugate(p::Partition{T}) where T<:Integer

returns the conjugate partition of $p$.

The conjugate of a partition $p$ is obtained by writing $p$ as a diagram and then flipping it along it's main diagonal.

For more information see: Wikipedia:Partition

JuLie.partitions — Function

partitions(n::Integer)

A list of all partitions of an integer $n ≥ 0$, produced in lexicographically descending order (like in SAGE, but opposite to GAP (you can apply reverse() to reverse the order)).

The algorithm used is the algorithm ZS1 by A. Zoghbi and I. Stojmenovic, "Fast algorithms for generating integer partitions", Int. J. Comput. Math. 70 (1998), no. 2, 319–332.

You can increase performance by casting $n$ into a smaller integer type, e.g.

partitions(Int8(90))

partitions(m::Integer, n::Integer, l1::Integer, l2::Integer; z=0)

All partitions of an integer $m ≥ 0$ into $n ≥ 0$ parts with lower bound $l1 ≥ 0$ and upper bound $l2 ≥ l1$. Parameter $z$ should be set to 0 for arbitrary choice of parts (default), 1 for distinct parts. The partitions are produced in decreasing order.

The algorithm used is "parta" by W. Riha and K. R. James, "Algorithm 29. Efficient Algorithms for Doubly and Multiply Restricted Partitions" (1976). De-gotoed from ALGOL code by Elisa!

partitions(m::Integer, n::Integer)

All partitions of an integer $m ≥ 0$ into $n ≥ 1$ parts (no further restrictions). This simply calls partitions(m,n,1,m,z=0).

JuLie.ascending_partitions — Function

ascending_partitions(n::Integer;alg="ks")

Instead of encoding a partition of an integer $n ≥ 0$ as a descending sequence (which is our convention), one can also encode it as an ascending sequence. In the papers below it is claimed that generating the list of all ascending partitions is more efficient than generating descending ones. To test this, I have implemented the algorithms:

"ks" (default) is the algorithm AccelAsc (Algorithm 4.1) by J. Kelleher and B. O'Sullivan, "Generating All Partitions: A Comparison Of Two Encodings", https://arxiv.org/pdf/0909.2331.pdf, May 2014.

"m" is Algorithm 6 by M. Merca, "Fast Algorithm for Generating Ascending Compositions", J. Math Model. Algor. (2012) 11:89–104. This is similar to "ks".

The ascending partitions are given here as arrays, not of type Partition since these are descending by convention.

I don't see a significant speed difference to the descending encoding:

julia> @btime partitions(Int8(90));
  3.376 s (56634200 allocations: 6.24 GiB)

julia> @btime ascending_partitions(Int8(90),alg="ks");
  3.395 s (56634200 allocations: 6.24 GiB)

julia> @btime ascending_partitions(Int8(90),alg="m");
  3.451 s (56634200 allocations: 6.24 GiB)

I am using "ks" as default since it looks slicker and I believe there is a tiny mistake in the publication of "m" (which I fixed).

Multipartitions

JuLie.Multipartition — Type

struct Multipartition{T} <: AbstractArray{Partition{T},1}

An $r$-component multipartition of an integer $n$ is an $r$-tuple of partitions $\lambda^{(1)},...,\lambda^{(r)}$ where each $\lambda^{(i)}$ is a partition of some $a_i$ and the $a_i$ sum to $n$.

You can create a multipartition with

Multipartition( [[2,1], [], [3,2,1]] )

As with partitions, you can cast into smaller integer types to increase performance:

Multipartition( Array{Int8,1}[[2,1], [], [3,2,1]] )

JuLie.multipartitions — Function

function multipartitions(n::T, r::Integer) where T<:Integer

A list of all $r$-component multipartitions of $n$. As for partitions, you can cast $n$ into smaller type for efficiency, e.g.

multipartitions(Int8(3),2)

Multiset_partitions

A Multiset_partition is a Multipartition $λ^{(1)},...,λ^{(r)}$ where each $\lambda^{(i)}$ is a non-empty partition.$\\$ Furthermore the ordering of the Partitions $\lambda^{(i)}$ is not important, i.e. $\lambda^{(1)},...,\lambda^{(r)}$=$\lambda^{(\sigma(1))},...,\lambda^{(\sigma(r))}$ $\\$ Note: This equality-relation is not implemented in this Module.

JuLie.multiset_partitions — Function

multiset_partitions(n::T) where T<:Integer

A list of all multiset_partitions of an integer $n ⋝ 0$.

The performance will suffer by casting $n$ into a smaller integer type, e.g.

multiset_partitions(Int8(20))

multiset_partitions(p::Partition{T})  where T<:Integer

A list of all possible multiset_partitions of a Partition, by regrouping its parts into Partitions.

The algorithm used is the algorithm M by , "The Art of Computer Programming - Volume 4A, Combinatotial Algorithms, Part 1" by Donald E. Knuth(2011), 429–430. De-gotoed, index-shifted and generalized.

multiset_partitions(n::T, r::Integer) where T<:Integer

A list of all multiset_partitions of an integer $n ⋝ 0$ into $r ⋝ 1$ parts.

multiset_partitions(p::Partition{T}, r::Integer) where T<:Integer

A list of all possible $r$-restricted multiset_partitions of a Partition, by regrouping its parts into Partitions.

The algorithm used is a version of the algorithm M by , "The Art of Computer Programming - Volume 4A, Combinatotial Algorithms, Part 1" by Donald E. Knuth, 429–430 http://www.cs.utsa.edu/~wagner/knuth/fasc3b.pdf. De-gotoed, index-shifted and generalized.

JuLie.partition_to_partcount — Function

partition_to_partcount(p::Partition{T})  where T<:Integer

returns the part-count representation of a partition $p$, where the $n$-th element is the count of appearances of $n$ in $p$.

julia> partition_to_partcount([5,3,3,3,2,1,1])
  5-element Array{Int64,1}:
  2
  1
  3
  0
  1

for performance, partitions with trailing zeroes will not be allowed.

JuLie.partcount_to_partition — Function

partcount_to_partition(pc::Array{T,1}) where T<:Integer

returns the partition from a part-count representation $pc$ of a partition.

julia> partcount_to_partition([2,0,1])
  [3,1,1]

Tableaux

JuLie.Tableau — Type

struct Tableau{T} <: AbstractArray{AbstractArray{T,1},1}

A tableau or Young tableau is a finite collection of boxes, each holding a value. The boxes are arranged in left-justified rows. These rows are read from top to bottom. The row lengths have to be non-increasing order. You can create a tableau with

Tab=Tableau([[1,2,3],[4,5],[6]])

and then work with it like with an array of arrays. In fact, Tableau is a subtype of AbstractArray{AbstractArray{T,1},1}. Note that for efficiency the Tableau constructor does not check whether the given array is in fact a Tableau, i.e. a sequence of Arrays with decreasing lenghts. That's your job. The implementation of a subtype of AbstractArray is explained in the Julia documentation.

Furthermore there are two spezial types of tableaux:

A semistandard tableaux is a tableaux where the values are non decreasing in each row, and strictly increasing in each column.

A standard tableaux is a tableaux where the values are strictly increasing in each row and column. A standard tableaux also requires each Integer from 1 to the amount of boxes to occur exactly once.

JuLie.shape — Function

shape(Tab::Tableau{T})

returns the shape of $Tab$ as a Partition

The shape of a Tableau describes the length of each row. i.e:

julia> shape([ [1,2,3,4], [5,6], [7] ])
  [4, 2, 1]

JuLie.weight — Function

weight(Tab::Tableau)

returns the weight of the tableau $Tab$.

i.e. $w$::Array{Int,1} such that $w[i]$ = number of appearances of $i$ in $Tab$

JuLie.reading_word — Function

reading_word(Tab::Tableau)

returns an Array containing the cells of $Tab$ read from left to right and from bottom to top.

julia> reading_word([ [1,2,3] , [4,5] , [6] ])
6-element Array{Int64,1}:
 6
 4
 5
 1
 2
 3

JuLie.hook_length — Function

hook_length(Tab::Tableau, i::Integer, j::Integer)

returns the hook length of $Tab[i][j]$.

The hook length of a cell, is the count of cells to the right in the same row + the count of cells below in the same column + 1

assumes that $Tab[i][j]$ exists

hook_length(s::Partition, i::Integer, j::Integer)

returns the hook length of $Tab[i][j]$ for a Tableau $Tab$ of shape $s$.

assumes that $i≤length(s)$ and $j≤s[i]$

JuLie.hook_length_formula — Function

hook_length_formula(s::Partition)

returns the hook length formula for a tableau of shape $s$.

\[f^{λ} = {\dfrac{n!}{∏ h_λ(i,j)}}\]

where the product is over all cells $(i,j)$ in $Tab$, and $h_λ$ is the hook_length.

Equals the number of standard tableaux of shape $s$

JuLie.schensted — Function

schensted(sigma::Array{Integer,1})

An implementation of the Schensted algorithm from the Robinson-Schensted correspondence. $sigma$ represents the second line of a Permutation in two-line notation: $1->sigma[1] , 2->sigma[2] ,…$ Returns a pair of standard tableaux $P,Q$ (insertion- and recording- tableaux)

JuLie.bump! — Function

bump!(Tab::Tableau, x::Int)

Inserts $x$ into $Tab$ according to the Bumping algorithm by applying the Schensted insertion.

bump!(Tab::Tableau, x::Integer, Q::Tableau, y::Integer)

Inserts $x$ into $Tab$ according to the Bumping algorithm by applying the Schensted insertion. Traces the change with $Q$ by inserting $y$ at the same Position in $Q$ as $x$ in $Tab$.

semistandard Tableaux

A semistandard tableaux is a tableaux where the values are non decreasing in each row, and strictly increasing in each column.

JuLie.semistandard_tableaux — Function

semistandard_tableaux(shape::Partition{T}, max_val=sum(shape)::Integer) where T<:Integer

returns an Array containing all semistandard tableaux of shape $shape$ and elements $≤ max\_val$. The tableaux are in lexicographic order from left to right and top to bottom.

semistandard_tableaux(shape::Array{T,1}, max_val=sum(shape)::Integer) where T<:Integer

returns an Array containing all semistandard tableaux of shape $shape$ and elements $≤ max\_val$. The tableaux are in lexicographic order from left to right and top to bottom

semistandard_tableaux(box_num::T, max_val=box_num::T) where T<:Integer

returns an Array containing all semistandard tableaux made from $box\_num$ boxes and elements $≤ max\_val$.

semistandard_tableaux(s::Array{T,1}, weight::Array{T,1}) where T<:Integer

returns an Array containing all semistandard tableaux with shape $s$ and weights $weight$.

requires that $sum(s) = sum(weight)$

semistandard_tableaux(s::Partition{T}, weight::Partition{T}) where T<:Integer

returns an Array containing all semistandard tableaux with shape $s$ and weights $weight$.

requires that $sum(s) = sum(weight)$

JuLie.is_semistandard — Function

is_semistandard(Tab::Tableau)

Checks if $Tab$ is a semistandard tableau. i.e. a Tableau with non decresing rows and strictly increasing columns.

standard Tableaux

A standard tableaux is a tableaux where the values are strictly increasing in each row and column. A standard tableaux also requires each Integer from 1 to the amount of boxes to occur exactly once.

JuLie.standard_tableaux — Function

standard_tableaux(s::Array{Integer,1})

returns a list of all standard tableaux of a given shape $s$

standard_tableaux(s::Partition)

returns a list of all standard tableaux of a given shape $s$

standard_tableaux(n::Integer)

returns a list of all standard tableaux of size $n$

JuLie.is_standard — Function

is_standard(Tab::Tableau)

Checks if $Tab$ is a standard tableau. i.e. a Tableau with strictly increasing columns and rows, where each number from 1 to n appears exactly once.

Kostka Polynomials

JuLie.kostka_polynomial — Function

kostka_polynomial(lambda::Partition{T}, mu::Partition{T})

returns the Kostka polynomial $K_{lambda,mu}$ as a fmpz_poly over ZZ in t.

The Kostka polynomials are a Family of polynomials $K_{λμ}$ relating Hall-Littlewood polynomials $P_μ$ to Schur polynomials $s_λ$ :

\[s_λ(x_1,…,x_n) =\sum_μ K_{λμ}(t)P_μ(x_1,…,x_n;t)\]

They can be formulated as:

\[K_{λμ}(t) = ∑_{T ∈ SSYT(λ,μ)} t^{charge(T)}\]

But I used the following fomula:

\[K_{λ,μ}(t)=∑_{\{v\}}∏_{K=1}^{l(λ)-1}∏_{n≥1} \begin{bmatrix} ℙ_n^{(K)}+m_n(v^{(K)})\\ m_n(v^{(K)}) \end{bmatrix}_t\]

where the sum is over all admissible configurations $\{v\}$ i.e. sequences of partitions $v^{(K)}$ with

\[v^{(0)}=μ \text{ ,\hspace{2mm} } |v^{(K)}|=∑_{j≥K+1}λ_j \text{\hspace{5mm} and \hspace{5mm}} ℙ_n^{(K)}≥0 \text{\hspace{2mm} for all \hspace{1mm}} n>0, K=0,1,…l(λ)\]

\[\begin{aligned} & ℙ_n^{(K)} := ∑_{j≥1}\left[min\left(n,v_j^{(K+1)}\right) - 2⋅min\left(n,v_j^{(K)}\right) + min\left(n,v_j^{(K-1)}\right)\right] \\ & c(v) := ∑_{i≥1}(i-1)μ_i + ∑_{K=1}^{l(λ)-1}\left(𝕄\left[v^{(K)}, v^{(K)}\right] - 𝕄\left[v^{(K)}, v^{(K-1)}\right]\right) \\ & 𝕄[ρ,κ] := ∑_{i,j≥1} min(ρ_i,κ_j) \end{aligned}\]

$\left[\genfrac{[}{]}{0pt}{0}{m}{n} \right]_t$ the Gaussian binomial coefficient

The Algorithm is based on A matrix model for WZW (Apendix B) by Nick Dorey, David Tonga and Carl Turner

kostka_polynomial(lambda::Array{Integer,1}, mu::Array{Integer,1})

Shortcut for kostka_polynomial.

Schur Polynomials

JuLie.schur_polynomial — Function

schur_polynomial(shp::Partition)

returns the Schur function $s_{shp}$ as a Multivariate Polynomial.

\[s_{shp}:=∑_{T} x_1^{m_1}…x_n^{m_n}\]

where the sum is taken over all semistandard tableaux $T$ of shape $shp$, and $m_i$ gives the weight of $i$ in $T$.

Enumerative functions

JuLie.num_partitions — Function

num_partitions(n::fmpz)
num_partitions(n::Integer)

The number of integer partitions of the integer $n ≥ 0$. Uses the function from FLINT, which is really fast.

For more information on these numbers, see OEIS.

num_partitions(n::fmpz, k::fmpz)
num_partitions(n::Integer, k::Integer)

The number of integer partitions of the integer $n ≥ 0$ into $k ≥ 0$ parts. The implementation uses a recurrence relation.

For more information on these numbers, see OEIS.

JuLie.catalan — Function

catalan(n::fmpz; alg="binomial")
catalan(n::Integer; alg="binomial")

The n-th Catalan number. This counts a gazillion of things, see OEIS for more information. There are two algorithms implemented:

"binomial" (default): uses a simple formula with binomial coefficients.

"iterative": uses an iterative computation.

The binomial computation is much faster:

julia> @time x=catalan( ZZ(10)^5 , alg="binomial");
 0.007727 seconds (9 allocations: 95.750 KiB)

julia> @time x=catalan( ZZ(10)^5 , alg="iterative");
 1.572488 seconds (1.01 M allocations: 2.333 GiB, 1.36% gc time)

JuLie.stirling1 — Function

stirling1(n::fmpz, k::fmpz)
stirling1(n::Integer, k::Integer)

The Stirling number $S_1(n,k)$ of the first kind. The absolute value of $S_1(n,k)$ counts the number of permutations of $n$ elements with $k$ disjoint cycles. The implementation is a wrapper to the function in FLINT.

For more information on these numbers, see OEIS.

JuLie.stirling2 — Function

stirling2(n::fmpz, k::fmpz)
stirling2(n::Integer, k::Integer)

The Stirling number $S_2(n,k)$ of the second kind. This counts the number of partitions of an $n$-set into $k$ non-empty subsets. The implementation is a wrapper to the function in FLINT.

For more information on these numbers, see OEIS.

JuLie.lucas — Function

lucas(n::fmpz)
lucas(n::Integer)

The n-th Lucas number. For more information on these numbers, see OEIS. The implementation is a wrapper to the function in GMP.

Quantum numbers

JuLie.quantum_number — Function

quantum_number(n::Int, q::RingElem)

For an integer $n \geq 0$ and a ring element $q$ which is invertible in its parent ring, the quantum integer $[n]_q$ is for $n \geq 0$ defined as $[n]_q = \sum_{i=0}^{n-1} q^{n-(2i+1)}$ and for $n < 0$ as $[n]_q = -[-n]_q$.

quantum_number(n::Int)

The quantum number $[n]_q$ where q is the interdeterminate of the Laurent polynomial ring $\mathbb{Z}[q,q^{-1}]$ in one variable over the integers.

JuLie.quantum — Function

quantum(n::Int, q::RingElem)

This is a shortcut for quanum_number(n,q).

quantum(n::Int)

This is a shortcut for quanum_number(n).

Index

JuLie.Multipartition

JuLie.Partition

JuLie.Tableau

JuLie.ascending_partitions

JuLie.bump!

JuLie.catalan

JuLie.conjugate

JuLie.dominates

JuLie.getelement

JuLie.hook_length

JuLie.hook_length_formula

JuLie.is_semistandard

JuLie.is_standard

JuLie.kostka_polynomial

JuLie.lucas

JuLie.multipartitions

JuLie.multiset_partitions

JuLie.num_partitions

JuLie.partcount_to_partition

JuLie.partition_to_partcount

JuLie.partitions

JuLie.quantum

JuLie.quantum_number

JuLie.reading_word

JuLie.schensted

JuLie.schur_polynomial

JuLie.semistandard_tableaux

JuLie.shape

JuLie.standard_tableaux

JuLie.stirling1

JuLie.stirling2

JuLie.weight