Partitions and the like

Compositions

JuLie.Composition — Type

Composition{T} <: AbstractArray{T,1}

A composition of an integer n ≥ 0 is a sequence (λ₁,…,λᵣ) of positive integers whose sum is equal to n. We have implemented an own type Composition as subtype of AbstractArray{T,1}.

Example

julia> c=Composition([2,1])
julia> sum(c)
3

JuLie.compositions — Function

compositions(n::Integer, k::Integer)

Returns an array of all compositions of n into k parts. The algorithm used is Algorithm 72 "Composition Generator" by L. Hellerman and S. Ogden, Communications of the ACM, 1961 (which refers to John Riordan, An Introduction to Combinatorial Analysis, 1958, Chapter 6). There may be faster algorithms—I don't know—but we're at least way faster than Sage (and also GAP4). De-gotoed by Elisa (as usual).

julia> @time X=compositions(26);
  5.189374 seconds (33.56 M allocations: 6.863 GiB, 20.09% gc time)

compositions(n::Integer)

Returns an array of all compositions of an integer n. This iterates over compositions of n into k parts for 1 ≤ k ≤ n.

JuLie.num_compositions — Function

num_compositions(n::Integer)

The number of compositons of an integer n>0 is equal to $2^{n-1}$. For n=0 there is 1 composition (the empty one) by convention. See OEIS for more information.

Partitions

JuLie.Partition — Type

Partition{T} <: AbstractArray{T,1}

A partition of an integer n ≥ 0 is a decreasing sequence λ=(λ₁,…,λᵣ) of positive integers λᵢ whose sum is equal to n. The λᵢ are called the parts of the partition. We encode a partition as an array with elements λᵢ. To be able to conceptually work with partitions we have implemented an own type Partition{T} as subtype of AbstractArray{T,1}. All functions for arrays then also work for partitions. You may increase performance by using smaller integer types, see the example below. For efficiency, the Partition constructor does not check whether the given array is in fact a partition, i.e. a decreasing sequence.

For more general information on partitions, check out Wikipedia.

Example

julia> P=Partition([3,2,1]) #The partition 3+2+1 of 6
julia> sum(P) #The sum of the parts.
6
julia> P[1] #First component
3
julia> P=Partition(Int8[3,2,1]) #Same partition but using 8 bit integers

Remarks

Usually, |λ| ≔ n is called the size of λ. In Julia, the function size for arrays already exists and returns the dimension of an array. Instead, you can use sum to get the sum of the parts.

There is no performance impact by using an own type for partitions rather than simply using arrays—I've tested this. Julia is great. The implementation of a subtype of AbstractArray is explained in the Julia documentation.

JuLie.partitions — Function

partitions(n::Integer)

A list of all partitions of an integer n ≥ 0, produced in lexicographically descending order. This ordering is like in SAGE, but opposite to GAP. You can apply reverse() to reverse the order. As usual, you may increase performance by using smaller integer types. The algorithm used is [1, Algorithm ZS1].

Example

julia> partitions(Int8(90)) #Using 8-bit integers

References

A. Zoghbi and I. Stojmenovic, "Fast algorithms for generating integer partitions", Int. J. Comput. Math. 70 (1998), no. 2, 319–332.

partitions(m::Integer, n::Integer, l1::Integer, l2::Integer; z=0)

A list of all partitions of an integer m ≥ 0 into n ≥ 0 parts with lower bound l1 ≥ 0 and upper bound l2 ≥ l1 for the parts. There are two choices for the parameter z:

z=0: no further restriction (default);

z=1: only distinct parts.

The partitions are produced in decreasing order.

The algorithm used is "parta" in 1.

Refereces

W. Riha and K. R. James, "Algorithm 29. Efficient Algorithms for Doubly and Multiply Restricted Partitions" (1976).

partitions(m::Integer, n::Integer)

All partitions of an integer m ≥ 0 into n ≥ 1 parts (no further restrictions). This simply calls partitions(m,n,1,m,z=0).

partitions(mu::Array{Integer,1}, m::Integer, v::Array{Integer,1}, n::Integer)

All partitions of an integer m >= 0 into n >= 1 parts, where each part is an element in v and each v[i] occurs a maximum of mu[i] times. The partitions are produced in decreasing order. The algorithm used is a de-gotoed version of "partb" in [1].

Remark

The original algorithm lead to BoundsErrors, since r could get smaller than 1. Furthermore x and y are handled as Arrays with an infinite length. After finding all valid partitions, the algorithm will continue searching for partitions of length n+1. We thus had to add a few additional checks and interruptions.

References

W. Riha and K. R. James, "Algorithm 29. Efficient Algorithms for Doubly and Multiply Restricted Partitions" (1976).

JuLie.num_partitions — Function

num_partitions(n::fmpz)
num_partitions(n::Integer)

The number of integer partitions of the integer n ≥ 0. Uses the function from FLINT, which is really fast.

References

OEIS

num_partitions(n::fmpz, k::fmpz)
num_partitions(n::Integer, k::Integer)

The number of integer partitions of the integer n ≥ 0 into k ≥ 0 parts. The implementation uses a recurrence relation.

References

OEIS

JuLie.dominates — Function

dominates(λ::Partition, μ::Partition)

The dominance order on partitions is the partial order ⊵ defined by λ ⊵ μ if and only if λ₁ + … + λᵢ ≥ μ₁ + … + μᵢ for all i. This function returns true if λ ⊵ μ.

References

Wikipedia.

JuLie.conjugate — Function

conjugate(λ::Partition{T}) where T<:Integer

The conjugate of a partition is obtained by considering its Young diagram (see Tableau) and then flipping it along its main diagonal.

References

Wikipedia.

JuLie.getelement — Function

getelement(P::Partition, i::Int)

Sometimes in algorithms for partitions it is convenient to be able to access parts beyond the length of the partition, and then you want to get zero instead of an error. This function is a shortcut for

return (i>length(P.p) ? 0 : getindex(P.p,i))

If you are sure that P[i] exists, use getindex because this will be faster.

JuLie.ascending_partitions — Function

ascending_partitions(n::Integer;alg="ks")

Instead of encoding a partition of an integer n ≥ 0 as a descending sequence (which is our convention), one can also encode it as an ascending sequence. In the papers [1,2] it is said that generating the list of all ascending partitions is more efficient than generating descending ones. To test this, I have implemented the algorithms:

"ks" (default) is the algorithm AccelAsc (Algorithm 4.1) in [1].

"m" is Algorithm 6 in [2]. This is similar to "ks".

The ascending partitions are given here as arrays, not of type Partition since these are descending by our convention. I am using "ks" as default since it looks slicker and I believe there is a tiny mistake in the publication of "m" (which I fixed).

Comparison

I don't see a significant speed difference to the descending encoding:

julia> @btime partitions(Int8(90));
	3.376 s (56634200 allocations: 6.24 GiB)

julia> @btime ascending_partitions(Int8(90),alg="ks");
	3.395 s (56634200 allocations: 6.24 GiB)

julia> @btime ascending_partitions(Int8(90),alg="m");
	3.451 s (56634200 allocations: 6.24 GiB)

References

J. Kelleher and B. O'Sullivan, "Generating All Partitions: A Comparison Of Two Encodings", https://arxiv.org/pdf/0909.2331.pdf, May 2014.

Merca, "Fast Algorithm for Generating Ascending Compositions", J. Math Model. Algor. (2012) 11:89–104.

Multipartitions

JuLie.Multipartition — Type

Multipartition{T} <: AbstractArray{Partition{T},1}

Multipartitions are generalizations of partitions. An r-component multipartition of an integer n is an r-tuple of partitions λ¹, λ², …, λʳ where each λⁱ is a partition of some integer nᵢ ≥ 0 and the nᵢ sum to n. As for partitions, we have implemented an own type Multipartition{T} which is a subtype of AbstractArray{Partition{T},1}. As with partitions, you can can use smaller integer types to increase performance.

Example

julia> P=Multipartition( [[2,1], [], [3,2,1]] )
julia> sum(P)
9
julia> P[2]
Int64[]
julia> Multipartition( Array{Int8,1}[[2,1], [], [3,2,1]] ) #Using 8-bit integers

JuLie.multipartitions — Function

multipartitions(n::T, r::Integer) where T<:Integer

A list of all r-component multipartitions of n.

Example

julia> multipartitions(Int8(3),2) #Using 8-bit integers

JuLie.num_multipartitions — Function

num_multipartitions(n::Int, k::Int)

The number of multipartitions of n into k parts. There's a straightforward way to count them, see [1, Proof of Lemma 2.4].

References

D. Craven, The Number of t-Cores of Size n (2006).

Base.sum — Method

sum(P::Multipartition{T}) where T<:Integer

If P is a multipartition of the integer n, this function returns n.

Multiset partitions

JuLie.Multiset_partition — Type

Multiset_partition{T}

Multiset-partitions are generalizations of partitions. An r-component multiset-partition of an integer $n$ is a multiset(a set where each element can be contained multiple times) of partitions $λ¹, λ², …, λʳ$ where each $λⁱ$ is a partition of some integer $nᵢ ≥ 1$ and the $nᵢ$ sum to $n$. As for partitions, we have implemented an own type Multiset_partition{T}. As with partitions, you can can use smaller integer types to increase performance.

Example

julia> P = Multiset_partition( [2,1], [4], [3,2,1] )
{[2, 1], [4], [3, 2, 1]}
julia> sum(P)
13
julia> Multiset_partition( Array{Int8,1}[[2,1], [4], [3,2,1]] ) #Using 8-bit integers
{Int8[2, 1], Int8[4], Int8[3, 2, 1]}

Since Multiset-partitions are unordered sets, you can't call an explicit element, however, you can iterate over a Multiset_partition.

Example

julia> MSP = Multiset_partition( [2,1], [4], [3,2,1] )
{[2, 1], [4], [3, 2, 1]}
julia> for p in MSP println(p) end
[2, 1]
[4]
[3, 2, 1]

JuLie.multiset_partitions — Function

multiset_partitions(n::T) where T<:Integer

A list of all multiset_partitions of an integer $n ⋝ 0$.

The performance will suffer by casting $n$ into a smaller integer type, e.g.

multiset_partitions(Int8(20))

multiset_partitions(p::Partition{T})	where T<:Integer

A list of all possible multiset_partitions of a Partition, by regrouping its parts into Partitions.

The algorithm used is the algorithm M by , "The Art of Computer Programming - Volume 4A, Combinatorial Algorithms, Part 1" by Donald E. Knuth(2011), 429–430. De-gotoed, index-shifted and generalized.

multiset_partitions(n::T, r::Integer) where T<:Integer

A list of all multiset_partitions of an integer $n ⋝ 0$ into $r ⋝ 1$ parts.

multiset_partitions(p::Partition{T}, r::Integer) where T<:Integer

A list of all possible $r$-restricted multiset_partitions of a Partition, by regrouping its parts into Partitions.

The algorithm used is a version of the algorithm M by , "The Art of Computer Programming - Volume 4A, Combinatorial Algorithms, Part 1" by Donald E. Knuth, 429–430 http://www.cs.utsa.edu/~wagner/knuth/fasc3b.pdf. De-gotoed, index-shifted and generalized.

Tableaux

JuLie.Tableau — Type

Tableau{T} <: AbstractArray{AbstractArray{T,1},1}

A Young diagram is a diagram of finitely many empty "boxes" arranged in left-justified rows, with the row lengths in non-increasing order. The box in row i and and column j has the coordinates (i,j). Listing the number of boxes in each row gives a partition λ of a non-negative integer n (the total number of boxes of the diagram). The diagram is then said to be of shape λ. Conversely, one can associate to any partition λ a Young diagram in the obvious way, so Young diagrams are just another way to look at partitions.

A Young tableau of shape λ is a filling of the boxes of the Young diagram of λ with elements from some set. After relabeling we can (and will) assume that we fill from a set of integers from 1 up to some number, which in applications is often equal to n. We encode a tableau as an array of arrays and we have implemented an own type Tableau{T} as subtype of AbstractArray{AbstractArray{T,1},1} to work with tableaux. As for partitions, you may increase performance by casting into smaller integer types, e.g.

For efficiency, we do not check whether the given array is really a tableau, i.e. whether the structure of the array defines a partition.

For more information see Wikipedia.

Example

Tab=Tableau([[1,2,3],[4,5],[6]])
Tab=Tableau(Array{Int8,1}[[2,1], [], [3,2,1]]) #Using 8 bit integers

JuLie.shape — Function

shape(Tab::Tableau{T})

Returns the shape of a tableau.

JuLie.weight — Function

weight(Tab::Tableau)

The weight of a tableau is the number of times each number appears in the tableau. The return value is an array whose i-th element gives the number of times the integer i appears in the tableau.

JuLie.reading_word — Function

reading_word(Tab::Tableau)

The reading word of a tableau is the word obtained by concatenating the fillings of the rows, starting from the bottom row. The word is here returned as an array.

Example

julia> reading_word(Tableau([ [1,2,3] , [4,5] , [6] ]))
6-element Array{Int64,1}:
 6
 4
 5
 1
 2
 3

JuLie.is_semistandard — Function

is_semistandard(Tab::Tableau)

A tableau is called semistandard if the entries weakly increase along each row and strictly increase down each column.

JuLie.semistandard_tableaux — Function

semistandard_tableaux(shape::Partition{T}, max_val=sum(shape)::Integer) where T<:Integer

Returns a list of all semistandard tableaux of given shape and filling elements bounded by max_val. By default, max_val is equal to the sum of the shape partition (the number of boxes in the Young diagram). The list of tableaux is in lexicographic order from left to right and top to bottom.

semistandard_tableaux(shape::Partition{T}, max_val=sum(shape)::Integer) where T<:Integer

Shortcut for semistandard_tableaux(Partition(shape),max_val).

semistandard_tableaux(box_num::T, max_val=box_num::T) where T<:Integer

Returns a list of all semistandard tableaux consisting of box_num boxes and filling elements bounded by max_val.

semistandard_tableaux(s::Array{T,1}, weight::Array{T,1}) where T<:Integer

Returns a list of all semistandard tableaux with shape s and given weight. This requires that sum(s) = sum(weight).

semistandard_tableaux(s::Partition{T}, weight::Partition{T}) where T<:Integer

Same as for semistandard_tableaux(s::Array{T,1}, weight::Array{T,1}) where T<:Integer.

JuLie.is_standard — Function

is_standard(Tab::Tableau)

A tableau is called standard if it is semistandard and the entries are in bijection with 1,…n, where n is the number of boxes.

JuLie.standard_tableaux — Function

standard_tableaux(s::Partition)

Returns a list of all standard tableaux of a given shape.

standard_tableaux(s::Array{Integer,1})

Shortcut for standard_tableaux(Partition(s)).

standard_tableaux(n::Integer)

Returns a list of all standard tableaux with n boxes.

JuLie.hook_length — Function

hook_length(lambda::Partition, i::Integer, j::Integer)

Consider the Young diagram of a partition λ. The hook length of a box, is the number of boxes to the right in the same row + the number of boxes below in the same column + 1. The function returns the hook length of the box with coordinates (i,j). The functions assumes that the box exists.

hook_length(Tab::Tableau, i::Integer, j::Integer)

Shortcut for hook_length(shape(Tab),i,j).

JuLie.hook_lengths — Function

hook_lengths(lambda::Partition)

Returns the tableau of shape λ in which the entry at position (i,j) is equal to the hook length of the corresponding box.

JuLie.num_standard_tableaux — Function

num_standard_tableaux(lambda::Partition)

Returns the number $f^\lambda$ of standard tableaux of shape λ using the hook length formula

\[f^{\lambda} = \frac{n!}{\prod_{i,j} h_\lambda(i,j)} \;,\]

where the product is taken over all boxes in the Young diagram of $\lambda$ and $h_\lambda$ denotes the hook length of the box (i,j).

For my information, see Wikipedia.

JuLie.schensted — Function

schensted(sigma::Array{Integer,1})

The Robinson–Schensted correspondence is a bijection between permutations and pairs of standard Young tableaux of the same shape. For a permutation sigma (given as an array), this function performs the Schnested algorithm and returns the corresponding pair of standard tableaux (the insertion and recording tableaux).

For more information, see Wikipedia.

Example

julia> schensted([3,1,6,2,5,4])
(AbstractArray{Int64,1}[[1, 2, 4], [3, 5], [6]], AbstractArray{Int64,1}[[1, 3, 5], [2, 4], [6]])

JuLie.bump! — Function

bump!(Tab::Tableau, x::Int)

Inserts the integer x into the tableau Tab according to the bumping algorithm by applying the Schensted insertion.

bump!(Tab::Tableau, x::Integer, Q::Tableau, y::Integer)

Inserts x into Tab according to the bumping algorithm by applying the Schensted insertion. Traces the change with Q by inserting y at the same Position in Q as x in Tab.