JuLie.jl

A Julia package for Combinatorics

JuLie.jl

Using

Partitions

Multipartitions

Enumerative functions

Quantum numbers

Index

Using

To install the package, do the following In Julia:

using Pkg
Pkg.add(url="https://github.com/ulthiel/JuLie.jl")

Then you can start using the package as follows:

using JuLie
partitions(10)

You can get a list of exported functions using

names(JuLie)

You can get help for a function by putting a question mark in front, e.g.

?partitions

Partitions

JuLie.Partition — Type

struct Partition{T} <: AbstractArray{T,1}

A partition of an integer n >= 0 is a decreasing sequence n1, n2, ... of positive integers whose sum is equal to n. You can create a partition with

P=Partition([3,2,1])

and then work with it like with an array. In fact, Partition is a subtype of AbstractArray{T,1}. You can increase performance by using smaller integer types, e.g.

P=Partition(Int8[3,2,1])

Note that for efficiency the Partition constructor does not check whether the given array is in fact a partition, i.e. a decreasing sequence. That's your job.

I was thinking back and forth whether to implement an own structure for this because it's actually just an array of integers. But it makes sense since we have several functions just acting on partitons and it would be strange implementing them for arrays in general (where mostly they don't make sense). I was hesitating because I feared that an own structure for partitions will have a performance impact. But it does not! In my standard example creating the partitions of 90 there is really NO difference in runtime and memory consumption between using arrays and using an own structure.

The implementation of a subtype of AbstractArray is explained in https://docs.julialang.org/en/v1/manual/interfaces/#man-interface-array.

JuLie.partitions — Function

partitions(n::Integer)

A list of all partitions of an integer n >= 0, produced in lexicographically descending order (like in SAGE, but opposite to GAP (you can apply reverse() to reverse the order)). The algorithm used is the algorithm ZS1 by A. Zoghbi and I. Stojmenovic, "Fast algorithms for generating integer partitions", Int. J. Comput. Math. 70 (1998), no. 2, 319–332.

You can increase performance by casting n into a smaller integer type, e.g.

partitions(Int8(90))

partitions(m::Integer, n::Integer, l1::Integer, l2::Integer; z=0)

All partitions of an integer m >= 0 into n >= 0 parts with lower bound l1>=0 and upper bound l2>=l1. Parameter z should be set to 0 for arbitrary choice of parts (default), 1 for distinct parts. The partitions are produced in decreasing order.

The algorithm used is "parta" by W. Riha and K. R. James, "Algorithm 29. Efficient Algorithms for Doubly and Multiply Restricted Partitions" (1976). De-gotoed from ALGOL code by Elisa!

partitions(m::Integer, n::Integer)

All partitions of an integer m >= 0 into n >= 1 parts (no further restrictions). This simply calls partitions(m,n,1,m,z=0).

JuLie.ascending_partitions — Function

ascending_partitions(n::Integer;alg="ks")

Instead of encoding a partition of an integer n >= 0 as a descending sequence (which is our convention), one can also encode it as an ascending sequence. In the papers below it is claimed that generating the list of all ascending partitions is more efficient than generating descending ones. To test this, I have implemented the algorithms:

"ks" (default) is the algorithm AccelAsc (Algorithm 4.1) by J. Kelleher and B. O'Sullivan, "Generating All Partitions: A Comparison Of Two Encodings", https://arxiv.org/pdf/0909.2331.pdf, May 2014.

"m" is Algorithm 6 by M. Merca, "Fast Algorithm for Generating Ascending Compositions", J. Math Model. Algor. (2012) 11:89–104. This is similar to "ks".

The ascending partitions are given here as arrays, not of type Partition since these are descending by convention.

I don't see a significant speed difference to the descending encoding:

julia> @btime partitions(Int8(90));
  3.376 s (56634200 allocations: 6.24 GiB)

julia> @btime ascending_partitions(Int8(90),alg="ks");
  3.395 s (56634200 allocations: 6.24 GiB)

julia> @btime ascending_partitions(Int8(90),alg="m");
  3.451 s (56634200 allocations: 6.24 GiB)

I am using "ks" as default since it looks slicker and I believe there is a tiny mistake in the publication of "m" (which I fixed).

Multipartitions

Missing docstring.

Missing docstring for Multipartition. Check Documenter's build log for details.

Missing docstring.

Missing docstring for multipartitions. Check Documenter's build log for details.

Enumerative functions

JuLie.num_partitions — Function

num_partitions(n::fmpz)
num_partitions(n::Integer)

The number of integer partitions of the integer n >= 0. Uses the function from FLINT, which is really fast.

For more information on these numbers, see https://oeis.org/A000041.

num_partitions(n::fmpz, k::fmpz)
num_partitions(n::Integer, k::Integer)

The number of integer partitions of the integer n >= 0 into k >= 0 parts. The implementation uses a recurrence relation.

For more information on these numbers, see https://oeis.org/A008284.

JuLie.catalan — Function

catalan(n::fmpz; alg="binomial")
catalan(n::Integer; alg="binomial")

The n-th Catalan number. This counts a gazillion of things, see https://oeis.org/A000108 for more information. There are algorithms implemented:

"binomial" (default): uses a simple formula with binomial coefficients.

"iterative": uses an iterative computation.

The binomial computation is much faster:

julia> @time x=catalan( ZZ(10)^5 , alg="binomial");
 0.007727 seconds (9 allocations: 95.750 KiB)

julia> @time x=catalan( ZZ(10)^5 , alg="iterative");
 1.572488 seconds (1.01 M allocations: 2.333 GiB, 1.36% gc time)

JuLie.stirling1 — Function

stirling1(n::fmpz, k::fmpz)
stirling1(n::Integer, k::Integer)

The Stirling number $S_1(n,k)$ of the first kind. The absolute value of $S_1(n,k)$ counts the number of permutations of n elements with k disjoint cycles. The implementation is a wrapper to the function in FLINT.

For more information on these numbers, see https://oeis.org/A008275.

JuLie.stirling2 — Function

stirling2(n::fmpz, k::fmpz)
stirling2(n::Integer, k::Integer)

The Stirling number $S_2(n,k)$ of the second kind. This counts the number of partitions of an n-set into k non-empty subsets. The implementation is a wrapper to the function in FLINT.

For more information on these numbers, see https://oeis.org/A008277.

JuLie.lucas — Function

lucas(n::fmpz)
lucas(n::Integer)

The n-th Lucas number. For more information on these numbers, see https://oeis.org/A000032. The implementation is a wrapper to the function in GMP.

Quantum numbers

JuLie.quantum_number — Function

quantum_number(n::Int, q::RingElem)

For an integer n>=0 and a ring element q which is invertible in its parent ring the quantum integer $[n]_q$ is for n>=0 defined as $[n]_q = sum_{i=0}^{n-1} q^{n-(2i+1)}$ and for n < 0 as $[n]_q = -[-n]_q$.

quantum_number(n::Int)

The quantum number [n]_q where q is the interdeterminate of the Laurent polynomial ring Z[q,q^-1] in one variable over the integers.

JuLie.quantum — Function

quantum(n::Int, q::RingElem)

This is a shortcut for quanum_number(n,q)

quantum(n::Int)

This is a shortcut for quanum_number(n)

Index

JuLie.Partition

JuLie.ascending_partitions

JuLie.catalan

JuLie.lucas

JuLie.num_partitions

JuLie.partitions

JuLie.quantum

JuLie.quantum_number

JuLie.stirling1

JuLie.stirling2

link to func(x)

link to partitions