Lie Theory

Cartan matrices

JuLie.CartanMatrix — Type

struct CartanMatrix <: MatElem{fmpz}

A (generalized) Cartan matrix is a square matrix $C$ over the integers satisfying the following conditions:

\[\begin{aligned} & C_{i,i} = 2 & \\ & C_{i,j} ≤ 0 \quad \text{ for } \quad i≠j \\ & C_{i,j} = 0 \quad \Leftrightarrow \quad C_{j,i} = 0 \;. \end{aligned}\]

This definition is as in Carter (2005). Note that in Kac (1990) more generally matrices over the complex numbers are considered—but most of the literature just focuses on the integral case.

You can create a CartanMatrix with

CartanMatrix( "A4~" )	 #calls essentially CartanMatrix('A', 4, true)
CartanMatrix( "C6" )		#calls essentially CartanMatrix('C', 6)

which generates the CartanMatrix: $\\tilde{A_4}$ resp. $C_6$

If you would like to generate block diagonal matrices you can do so with

CartanMatrix(["A3~","G2","B12"])	#calls essentially CartanMatrix(['A','G','B'],[3,2,12],[true,false,false])

References

Carter, R. (2005). Lie Algebras of Finite and Affine Type. Cambridge University Press, Cambridge.

Kac, V. G. (1990). Infinite dimensional Lie algebras (Third edition). Cambridge University Press, Cambridge. (First edition published 1982)

JuLie.is_cartan_matrix — Function

is_cartan_matrix(C::CartanMatrix)

Returns true iff $C$ is a (generalized) Cartan matrix.

JuLie.is_square — Function

is_square(M::MatElem)

Returns true iff $M$ is a square matrix, i.e. the number of rows is equal to the number of columns. If true, the dimension is returned as well.

JuLie.adjacency_graph — Function

adjacency_graph(M::MatElem)

The adjacency graph of an $(n × n)$-matrix $M$ is the undirected graph with vertices $1,…,n$ and an edge from $i$ to $j$ if and only if $i \ne j$ and $M_{ij} ≠ 0$. The adjacency graph is returned as a SimpleGraph fom the LightGraphs package.

JuLie.permute — Function

permute(M::MatElem, σ::Perm)

Given an $(n × n)$-matrix $M$ and a permutation $σ ∈ S_n$, this function returns the matrix $M^\sigma$ with $(M^\sigma)_{ij} ≔ M_{\sigma(i) \sigma(j)}$, i.e. the rows and columns are permuted by $σ$.

JuLie.block_decomposition — Function

block_decomposition(M::MatElem)

Let $M$ be an $(n × n)$-matrix. A block decomposition of $M$ is a collection of square matrices $B_1,\ldots,B_r$ together with a permutation $σ ∈ S_n$ such that the matrix $M^\sigma$ obtained by applying the permutation $\sigma$ to the rows and columns of $M$ is the block diagonal sum of the $B_i$. Up to reordering of the $B_i$ and up to simulatenous row and column permutation of each $B_i$, a block decomposition is unique and corresponds to the connected components of the adjacency graph of $M$.

JuLie.is_indecomposable — Function

is_indecomposable(C::MatElem)

A square matrix $M$ is indecomposable if it has no non-trivial block decomposition, i.e. the adjacency graph of $M$ is connected.