Lie Theory

Cartan matrices

JuLie.CartanMatrix — Type

struct CartanMatrix{T} <: MatElem{T}

A (generalized) Cartan matrix $C$ is a square matrix over the integers, satisfying the following rules:

\[\begin{aligned} & C_{i,i} = 2 & \\ & C_{i,j} ≤ 0 \quad \text{ for } \quad i≠j \\ & C_{i,j} = 0 \quad \Leftrightarrow \quad C_{j,i} = 0 \end{aligned}\]

We implement them as MatElem from AbstractAlgebra.jl over fmpz.

You can create a CartanMatrix with

CartanMatrix( "A4~" )	 #calls essentially CartanMatrix('A', 4, true)
CartanMatrix( "C6" )		#calls essentially CartanMatrix('C', 6)

which generates the CartanMatrix: $\tilde{A_4}$ resp. $C_6$

If you would like to generate block diagonal matrices you can do so with

CartanMatrix(["A3~","G2","B12"])	#calls essentially CartanMatrix(['A','G','B'],[3,2,12],[true,false,false])

JuLie.is_cartan_matrix — Function

is_cartan_matrix(C::CartanMatrix)

returns true iff C is a (generalized) Cartan matrix.

Complex reflection groups

A reflection in a finite-dimensional complex vector space V is an automorphism s ∈ GL(V) whose fixed space Ker(id-s) is of codimension 1 in V (i.e. it is a hyperplane). A complex reflection group is a finite subgroup W ⊆ GL(V) for some finite-dimensional complex vector space V such that W is generated by reflections.

References

Wikipedia

Shephard, G. C. & Todd, J. A. (1954). Finite unitary reflection groups. Canad. J. Math., 6, 274–304.

Lehrer, G. I. & Taylor, D. E. (2009). Unitary reflection groups. Cambridge University Press, Cambridge.

<!–

Missing docstring.

Missing docstring for ComplexReflectionGroup. Check Documenter's build log for details.

Missing docstring.

Missing docstring for IrreducibleComplexReflectionGroup. Check Documenter's build log for details.

Missing docstring.

Missing docstring for coxeter_number. Check Documenter's build log for details.

–>

<!–

Imprimitive complex reflection groups

Missing docstring.

Missing docstring for ImprimitiveComplexReflectionGroup. Check Documenter's build log for details.

Missing docstring.

Missing docstring for order. Check Documenter's build log for details.

Missing docstring.

Missing docstring for rank. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ngens. Check Documenter's build log for details.

Missing docstring.

Missing docstring for is_wellgenerated. Check Documenter's build log for details.

Missing docstring.

Missing docstring for degrees. Check Documenter's build log for details.

Missing docstring.

Missing docstring for exponents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for codegrees. Check Documenter's build log for details.

Missing docstring.

Missing docstring for coexponents. Check Documenter's build log for details.

Missing docstring.

Missing docstring for num_reflections. Check Documenter's build log for details.

Missing docstring.

Missing docstring for num_classes_reflections. Check Documenter's build log for details.

Missing docstring.

Missing docstring for num_hyperplanes. Check Documenter's build log for details.

–>