Lie Theory

Cartan matrices

JuLie.CartanMatrix — Type

struct CartanMatrix{T} <: MatElem{T}

A (generalized) Cartan matrix $C$ is a square matrix over the integers, satisfying the following rules:

\[\begin{aligned} & C_{i,i} = 2 & \\ & C_{i,j} ≤ 0 \quad \text{ for } \quad i≠j \\ & C_{i,j} = 0 \quad \Leftrightarrow \quad C_{j,i} = 0`` \end{aligned}\]

We implement them as MatElem from AbstractAlgebra.jl over fmpz.

You can create a CartanMatrix with

CartanMatrix( "A4~" )     #calls essentially CartanMatrix('A', 4, true)
CartanMatrix( "C6" )      #calls essentially CartanMatrix('C', 6)

which generates the CartanMatrix: $\tilde{A_4}$ resp. $C_6$

If you would like to generate block diagonal matrices you can do so with

CartanMatrix(["A3~","G2","B12"])    #calls essentially CartanMatrix(['A','G','B'],[3,2,12],[true,false,false])

JuLie.is_cartan_matrix — Function

is_cartan_matrix(C::CartanMatrix)

returns true iff C is a (generalized) Cartan matrix.